// src/common/enums/TextureKeys.ts
export enum TextureKeys {
  BUTTON_SPIN = 'button_spin',
  BUTTON_SPIN_PRESED = 'button_spin_presed',
  BUTTON_AUTO_SPIN_ENABLE = 'button_auto_spin_enable',
  BUTTON_AUTO_SPIN_DISABLE = 'button_auto_spin_disable',
  BUTTON_FAST_MODE_ENABLE = 'button_fast_mode_enable',
  BUTTON_FAST_MODE_DISABLE = 'button_fast_mode_disable',
  BET_IMAGE = 'bet_image',
  BET_MINUS = 'bet_minus',
  BET_PLUS = 'bet_plus',
}